Restructured the folder
