+++ app/register/member/components/HodLogo.tsx
