+++ app/register/member/components/InformationChild.tsx
